parse home information into its own file
