--- src/components/NasaImage.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-
-const NasaImage = () => {
-  return (
-    <div style={{ color: 'white' }}>
-      NASA FILE
-    </div>
-  )
-}
-
-export default NasaImage;
